@@ -42,7 +42,7 @@
     "typescript": "^4.0.2"
   },
   "peerDependencies": {
-    "react": "^16.2.0"
+    "react": "^16.9"
   },
   "repository": {
     "type": "git",
