@@ -1,7 +1,7 @@
 import React from 'react'
 import styles from './MissingDocumentTypesMessage.css'
 
-export default function MissingDocumentTypesMessage(props) {
+export default function MissingDocumentTypesMessage() {
   return (
     <div className={styles.root}>
       <h2>Empty schema</h2>
