@@ -1,10 +1,16 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import {WithReferringDocuments} from 'part:@sanity/base/with-referring-documents'
 
-export default function enhanceWithReferringDocuments(Component) {
-  function EnhancedWithReferringDocuments(props) {
-    // eslint-disable-next-line react/no-multi-comp
+interface EnhancedWithReferringDocumentsProps {
+  published?: Record<string, any>
+}
+
+export default function enhanceWithReferringDocuments<ComponentProps>(
+  Component: React.ComponentType<ComponentProps>
+) {
+  function EnhancedWithReferringDocuments(
+    props: ComponentProps & EnhancedWithReferringDocumentsProps
+  ) {
     const renderChild = ({isLoading, referringDocuments}) => (
       <Component
         {...props}
@@ -12,6 +18,7 @@ export default function enhanceWithReferringDocuments(Component) {
         isCheckingReferringDocuments={isLoading}
       />
     )
+
     return props.published ? (
       <WithReferringDocuments id={props.published._id}>{renderChild}</WithReferringDocuments>
     ) : (
@@ -21,8 +28,6 @@ export default function enhanceWithReferringDocuments(Component) {
 
   EnhancedWithReferringDocuments.displayName = `enhanceWithReferringDocuments(${Component.displayName ||
     Component.name})`
-  EnhancedWithReferringDocuments.propTypes = {
-    published: PropTypes.object
-  }
+
   return EnhancedWithReferringDocuments
 }
