@@ -1,13 +1,21 @@
-import PropTypes from 'prop-types'
+import {DefaultGridListItem} from '@sanity/base/__legacy/components'
 import React, {useContext} from 'react'
-import {Item as GridListItem} from 'part:@sanity/components/lists/grid'
 import {PaneRouterContext} from '../../contexts/PaneRouterContext'
-import styles from './PaneItem.css'
 import listStyles from '../../components/listView/ListView.css'
+import styles from './PaneItem.css'
 
-const PaneItemWrapper = props => {
+interface PaneItemWrapperProps {
+  id: string
+  layout?: string
+  useGrid?: boolean
+  isSelected?: boolean
+  children?: React.ReactNode
+}
+
+const PaneItemWrapper = (props: PaneItemWrapperProps) => {
   const {ChildLink} = useContext(PaneRouterContext)
   const {id, useGrid, layout, isSelected} = props
+
   const link = (
     <ChildLink childId={id} className={isSelected ? styles.linkIsSelected : styles.link}>
       {props.children}
@@ -15,18 +23,10 @@ const PaneItemWrapper = props => {
   )
 
   return useGrid ? (
-    <GridListItem className={listStyles[`${layout}ListItem`]}>{link}</GridListItem>
+    <DefaultGridListItem className={listStyles[`${layout}ListItem`]}>{link}</DefaultGridListItem>
   ) : (
     <div className={isSelected ? styles.selected : styles.item}>{link}</div>
   )
-}
-
-PaneItemWrapper.propTypes = {
-  id: PropTypes.string.isRequired,
-  layout: PropTypes.string,
-  useGrid: PropTypes.bool,
-  isSelected: PropTypes.bool,
-  children: PropTypes.node
 }
 
 export default PaneItemWrapper
