@@ -1,8 +1,6 @@
-import {MenuItemGroup} from '@sanity/base/__legacy/components'
-import {Card} from '@sanity/ui'
+import {MenuItemGroup, ScrollContainer} from '@sanity/base/__legacy/components'
+import {Card, PortalProvider, usePortal} from '@sanity/ui'
 import classNames from 'classnames'
-import {PortalProvider, usePortal} from 'part:@sanity/components/portal'
-import {ScrollContainer} from 'part:@sanity/components/scroll'
 import React, {createElement, useCallback, useMemo, useRef} from 'react'
 import {Path} from '@sanity/types'
 import {useDeskToolFeatures} from '../../../features'
