@@ -1,17 +1,16 @@
-import {MenuItem, MenuItemGroup} from '@sanity/base/__legacy/components'
+import {Button, Menu, MenuButton, MenuItem} from '@sanity/ui'
+import {MenuItem as MenuItemType, MenuItemGroup} from '@sanity/base/__legacy/components'
 import IconMoreVert from 'part:@sanity/base/more-vert-icon'
-import {MenuButton} from 'part:@sanity/components/menu-button'
-import Menu from 'part:@sanity/components/menus/default'
-import React, {useCallback, useMemo} from 'react'
+import React, {useMemo} from 'react'
 
 import styles from './contextMenu.css'
 
 interface DocumentPanelContextMenuProps {
   boundaryElement: HTMLDivElement | null
   isCollapsed: boolean
-  items: MenuItem[]
+  items: MenuItemType[]
   itemGroups: MenuItemGroup[]
-  onAction: (action: MenuItem) => void
+  onAction: (action: MenuItemType) => void
   open: boolean
   setOpen: (val: boolean) => void
 }
@@ -27,45 +26,62 @@ export function DocumentPanelContextMenu(props: DocumentPanelContextMenuProps) {
     []
   )
 
-  const handleAction = useCallback(
-    (action: MenuItem) => {
-      onAction(action)
-      setOpen(false)
-    },
-    [onAction, setOpen]
-  )
+  // const handleAction = useCallback(
+  //   (action: MenuItemType) => {
+  //     onAction(action)
+  //     setOpen(false)
+  //   },
+  //   [onAction, setOpen]
+  // )
 
-  const handleCloseMenu = useCallback(() => {
-    setOpen(false)
-  }, [setOpen])
+  // const handleCloseMenu = useCallback(() => {
+  //   setOpen(false)
+  // }, [setOpen])
+
+  // console.log(items, itemGroups)
 
   return (
     <MenuButton
       boundaryElement={boundaryElement || undefined}
-      buttonProps={{
-        'aria-label': 'Menu',
-        'aria-haspopup': 'menu',
-        'aria-expanded': open,
-        'aria-controls': id,
-        className: styles.menuOverflowButton,
-        icon: IconMoreVert,
-        kind: 'simple',
-        padding: 'small',
-        selected: open,
-        title: 'Show menu'
-      }}
+      button={
+        <Button
+          aria-label="Menu"
+          aria-haspopup="menu"
+          aria-expanded={open}
+          aria-controls={id}
+          className={styles.menuOverflowButton}
+          icon={IconMoreVert}
+          // kind: 'simple',
+          // padding: 'small',
+          mode="bleed"
+          padding={2}
+          selected={open}
+          title="Show menu"
+        />
+      }
+      id="context-menu"
       menu={
         <Menu
           id={id}
-          items={items}
-          groups={itemGroups}
-          onAction={handleAction}
-          onClose={handleCloseMenu}
-        />
+          // items={items}
+          // groups={itemGroups}
+          // onAction={handleAction}
+          // onClose={handleCloseMenu}
+        >
+          {/* @todo: use item groups */}
+          {items.map((item, itemIndex) => (
+            <MenuItem
+              // icon={item.icon}
+              key={item.key || itemIndex}
+              onClick={() => onAction(item)}
+              text={item.title}
+            />
+          ))}
+        </Menu>
       }
-      open={open}
+      // open={open}
       placement="bottom"
-      setOpen={setOpen}
+      // setOpen={setOpen}
     />
   )
 }
