@@ -3,7 +3,7 @@
 import React from 'react'
 import {Subscription} from 'rxjs'
 import {isActionEnabled} from 'part:@sanity/base/util/document-action-utils'
-import Button from 'part:@sanity/components/buttons/default'
+import {Button} from '@sanity/ui'
 import schema from 'part:@sanity/base/schema'
 import afterEditorComponents from 'all:part:@sanity/desk-tool/after-editor-component'
 import filterFieldFn$ from 'part:@sanity/desk-tool/filter-fields-fn?'
