@@ -1,10 +1,7 @@
 import {useId} from '@reach/auto-id'
-import Button from 'part:@sanity/components/buttons/default'
-import {Popover} from 'part:@sanity/components/popover'
-import Hotkeys from 'part:@sanity/components/typography/hotkeys'
+import {Button, Hotkeys, Popover, useClickOutside} from '@sanity/ui'
 import ChevronDownIcon from 'part:@sanity/base/chevron-down-icon'
 import React, {createElement, useCallback, useEffect, useRef, useState} from 'react'
-import useOnClickOutside from 'use-onclickoutside'
 import {ActionStateDialog} from './actionStateDialog'
 
 import styles from './actionMenu.css'
@@ -25,9 +22,9 @@ interface Props {
 }
 
 export function ActionMenu({actionStates, onOpen, onClose, disabled, isOpen}: Props) {
-  const clickOutsideRef = useRef(null)
   const listRef = useRef<HTMLUListElement>(null)
   const idPrefix = useId()
+  const [rootElement, setRootElement] = useState<HTMLDivElement | null>(null)
 
   const handleCloseMenu = useCallback(() => {
     if (!isOpen) {
@@ -42,7 +39,7 @@ export function ActionMenu({actionStates, onOpen, onClose, disabled, isOpen}: Pr
     }
   }, [actionStates, isOpen, onClose])
 
-  useOnClickOutside(clickOutsideRef, handleCloseMenu)
+  useClickOutside(handleCloseMenu, [rootElement])
 
   const [activeAction, setActiveAction] = useState(actionStates.find(s => !s.disabled))
 
@@ -102,7 +99,7 @@ export function ActionMenu({actionStates, onOpen, onClose, disabled, isOpen}: Pr
   )
 
   return (
-    <div className={styles.actionsDropDown} ref={clickOutsideRef} onKeyDown={handleKeyDown}>
+    <div className={styles.actionsDropDown} ref={setRootElement} onKeyDown={handleKeyDown}>
       <Popover content={popoverContent} open={isOpen} placement="top-end">
         <div>
           <Button
@@ -112,7 +109,7 @@ export function ActionMenu({actionStates, onOpen, onClose, disabled, isOpen}: Pr
             disabled={disabled}
             icon={ChevronDownIcon}
             id={`${idPrefix}-button`}
-            kind="secondary"
+            mode="ghost"
             onClick={isOpen ? onClose : onOpen}
           />
         </div>
@@ -145,7 +142,7 @@ function ActionMenuListItem({actionState, activeAction, disabled}) {
           <span className={styles.menuItemLabel}>{actionState.label}</span>
           {actionState.shortcut && (
             <span className={styles.menuItemHotkeys}>
-              <Hotkeys keys={String(actionState.shortcut).split('+')} size="small" />
+              <Hotkeys keys={String(actionState.shortcut).split('+')} size={1} />
             </span>
           )}
         </div>
