@@ -15,14 +15,14 @@ export interface DocumentView {
   icon?: ComponentType<any>
 }
 
-export interface Doc {
-  _id?: string
-  _type?: string
-  _rev?: string
-  _updatedAt?: string
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [key: string]: any
-}
+// export interface Doc {
+//   _id?: string
+//   _type?: string
+//   _rev?: string
+//   _updatedAt?: string
+//   // eslint-disable-next-line @typescript-eslint/no-explicit-any
+//   [key: string]: any
+// }
 
 // export interface MenuAction {
 //   action: string
