@@ -1,18 +1,17 @@
 import React from 'react'
-import PropTypes from 'prop-types'
-import DefaultPane from 'part:@sanity/components/panes/default'
+import {Pane} from '../../components/pane'
 import styles from './UnknownPaneType.css'
 
-export default class UnknownPaneType extends React.PureComponent {
-  static propTypes = {
-    type: PropTypes.string,
-    isSelected: PropTypes.bool.isRequired,
-    isCollapsed: PropTypes.bool.isRequired,
-    onExpand: PropTypes.func,
-    onCollapse: PropTypes.func,
-    index: PropTypes.number
-  }
+interface UnknownPaneTypeProps {
+  type?: string
+  isSelected: boolean
+  isCollapsed: boolean
+  onExpand?: () => void
+  onCollapse?: () => void
+  index?: number
+}
 
+export default class UnknownPaneType extends React.PureComponent<UnknownPaneTypeProps> {
   static defaultProps = {
     type: undefined,
     onExpand: undefined,
@@ -23,7 +22,7 @@ export default class UnknownPaneType extends React.PureComponent {
     const {isSelected, isCollapsed, onCollapse, onExpand, type} = this.props
 
     return (
-      <DefaultPane
+      <Pane
         title="Unknown pane type"
         index={this.props.index}
         isSelected={isSelected}
@@ -44,7 +43,7 @@ export default class UnknownPaneType extends React.PureComponent {
             )}
           </p>
         </div>
-      </DefaultPane>
+      </Pane>
     )
   }
 }
