@@ -1,6 +1,4 @@
-/* eslint-disable @typescript-eslint/explicit-function-return-type */
-/* eslint-disable react/jsx-filename-extension */
-
+import {MasterDetailIcon} from '@sanity/icons'
 import React, {useEffect} from 'react'
 import UUID from '@sanity/uuid'
 import {route, useRouterState} from 'part:@sanity/base/router'
@@ -74,33 +72,6 @@ function getIntentState(intentName, params, currentState, payload) {
   }
 
   return {intent: intentName, params, payload}
-}
-
-const strokeStyle = {
-  stroke: 'currentColor',
-  strokeWidth: 1.2
-}
-
-// @todo: Move to @sanity/base
-// eslint-disable-next-line react/no-multi-comp
-function MasterDetailIcon() {
-  return (
-    <svg
-      data-sanity-icon
-      width="1em"
-      height="1em"
-      viewBox="0 0 25 25"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect x="4.5" y="5.5" width="16" height="14" style={strokeStyle} />
-      <path d="M10.5 5.5V19.5" style={strokeStyle} />
-      <path d="M6 8.5H9" style={strokeStyle} />
-      <path d="M13 8.5H18" style={strokeStyle} />
-      <path d="M6 11.5H9" style={strokeStyle} />
-      <path d="M6 14.5H9" style={strokeStyle} />
-    </svg>
-  )
 }
 
 function DeskTool() {
