@@ -1,20 +1,19 @@
 import * as React from 'react'
-import * as PropTypes from 'prop-types'
 
-export default class UseState extends React.Component {
-  constructor(props) {
+interface UseStateProps {
+  startWith?: any
+  children: (args: [any, (value: any) => void]) => React.ReactNode
+}
+
+export default class UseState extends React.Component<UseStateProps, {value: any}> {
+  constructor(props: UseStateProps) {
     super(props)
     this.state = {value: props.startWith}
   }
 
-  setValue = nextValue => this.setState({value: nextValue})
+  setValue = (nextValue: any) => this.setState({value: nextValue})
 
   render() {
     return this.props.children([this.state.value, this.setValue])
   }
-}
-
-UseState.propTypes = {
-  startWith: PropTypes.any,
-  children: PropTypes.func
 }
