@@ -1,14 +1,15 @@
 import {useEffect, useState} from 'react'
 import shallowEquals from 'shallow-equals'
-import {Observable, defer, throwError, from, of as observableOf} from 'rxjs'
+import {Observable, defer, throwError, from, of as observableOf, Subscription} from 'rxjs'
 import {map, switchMap, distinctUntilChanged} from 'rxjs/operators'
 import leven from 'leven'
+import generateHelpUrl from '@sanity/generate-help-url'
 import {LOADING_PANE} from '../constants'
 import defaultStructure from '../defaultStructure'
+import {StructureError} from '../types'
 import isSubscribable from './isSubscribable'
 import validateStructure from './validateStructure'
 import serializeStructure from './serializeStructure'
-import generateHelpUrl from '@sanity/generate-help-url'
 
 const KNOWN_STRUCTURE_EXPORTS = ['getDefaultDocumentNode']
 
@@ -46,7 +47,13 @@ function sumPaneSegments(paneGroups) {
   return paneGroups.reduce((count, curr) => count + curr.length, 0)
 }
 
-function resolveForStructure(structure, paneGroups, prevStructure, fromIndex, options = {}) {
+function resolveForStructure(
+  structure,
+  paneGroups,
+  prevStructure,
+  fromIndex,
+  options: {silent?: boolean} = {}
+) {
   return Observable.create(subscriber => {
     try {
       validateStructure(structure)
@@ -62,7 +69,7 @@ function resolveForStructure(structure, paneGroups, prevStructure, fromIndex, op
     const totalPanes = sumPaneSegments(paneSegments)
     const [fromRootIndex, fromSplitIndex] = fromIndex
     let panes = getInitialPanes(prevStructure, totalPanes, fromRootIndex + 1 + fromSplitIndex)
-    const subscriptions = []
+    const subscriptions: Subscription[] = []
 
     // Start with all-loading (or previous structure) state
     subscriber.next(panes)
@@ -110,7 +117,7 @@ function resolveForStructure(structure, paneGroups, prevStructure, fromIndex, op
     }
 
     function subscribeForUpdates(pane, index, splitIndex, context, resolverArgs) {
-      const source = serializeStructure(pane, context, resolverArgs)
+      const source: Observable<any> = serializeStructure(pane, context, resolverArgs)
       subscriptions.push(
         source.subscribe(
           result => emit(result, index, splitIndex),
@@ -206,7 +213,8 @@ function resolveForStructure(structure, paneGroups, prevStructure, fromIndex, op
 
     function unsubscribe() {
       while (subscriptions.length) {
-        subscriptions.pop().unsubscribe()
+        const sub = subscriptions.pop()
+        if (sub) sub.unsubscribe()
       }
     }
   })
@@ -255,13 +263,19 @@ export const loadStructure = () => {
   }
 
   // Defer to catch immediately thrown errors on serialization
-  return defer(() => serializeStructure(structure))
+  return defer(() => serializeStructure(structure, undefined))
 }
 
 export const useStructure = (segments, options = {}) => {
   const hasSegments = Boolean(segments)
   const numSegments = sumPaneSegments(segments || [])
-  const [{structure, error}, setStructure] = useState({structure: null, error: null})
+  const [{structure, error}, setStructure] = useState<{
+    structure?: any
+    error?: StructureError | null
+  }>({
+    structure: null,
+    error: null
+  })
 
   // @todo This leads to deep update loops unless we serialize paneSegments
   // @todo We should try to memoize/prevent this without resorting to JSON.stringify
@@ -314,7 +328,7 @@ function warnOnUnknownExports(mod) {
     .filter(key => !known.includes(key))
     .forEach(key => {
       const {closest} = known.reduce(
-        (acc, current) => {
+        (acc: {closest: any; distance: number}, current) => {
           const distance = leven(current, key)
           return distance < 3 && distance < acc.distance ? {closest: current, distance} : acc
         },
