@@ -1,4 +1,3 @@
-/* eslint-disable max-nested-callbacks */
 import React from 'react'
 import {streamingComponent} from 'react-props-stream'
 import {merge, from, of} from 'rxjs'
@@ -14,16 +13,16 @@ import {
 import schema from 'part:@sanity/base/schema'
 import {observePaths} from 'part:@sanity/base/preview'
 import {getDraftId, getPublishedId} from 'part:@sanity/base/util/draft-utils'
-import {PaneRouterContext} from '../contexts/PaneRouterContext'
-import {ErrorPane} from '../panes/errorPane'
-import {LoadingPane} from '../panes/loadingPane'
-import BrokenReferences from '../components/BrokenReferences'
 import {
   templateExists,
   getTemplateById,
   getTemplatesBySchemaType,
   resolveInitialValue
 } from '@sanity/base/initial-value-templates'
+import {PaneRouterContext} from '../contexts/PaneRouterContext'
+import {ErrorPane} from '../panes/errorPane'
+import {LoadingPane} from '../panes/loadingPane'
+import BrokenReferences from '../components/BrokenReferences'
 
 const withInitialValue = Pane => {
   const WithInitialValueStream = streamingComponent(props$ =>
