@@ -1,6 +1,1 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-exports.PreviewFields = require('./lib/components/PreviewFields').default
-exports.SanityPreview = require('./lib/components/SanityPreview').default
-exports.PreviewSubscriber = require('./lib/components/PreviewSubscriber').default
-exports.WithVisibility = require('./lib/components/WithVisibility').default
-exports.SanityDefaultPreview = require('./lib/components/SanityDefaultPreview').default
+module.exports = require('./lib/exports/components')
