@@ -1,15 +1,15 @@
 import React from 'react'
+import {Type} from '../types'
 import PreviewSubscriber from './PreviewSubscriber'
 import RenderPreviewSnapshot from './RenderPreviewSnapshot'
-import {Type} from '../types'
 
 interface Props {
   type: Type
-  fields: string[]
+  fields?: string[]
   value: any
-  ordering: {}
-  children: (props: any) => React.ComponentType
-  layout: string
+  ordering?: {}
+  children?: (props: any) => React.ComponentType
+  layout?: string
 }
 
 export default function SanityPreview(props: Props) {
